Move note list test to its own file
